hadk.config: Allow executing of chains before or after loading
Previously chains were only executed after the current file is loaded.
This resultet in the situation the execution was delayed until the
last dependency of it was loaded.
Now its possible to execute the chain right on the spot when its
loaded so they are executed in the proper order how depend is called.