templates.mer.droid.syspart: Reworked, cleaned up
Add separete variable $HYBRIS_ANDROID_SYSPART_BRANCH to specific a
branch for the syspart side of hybris-hal.

Also only initialise the linux parts of the manifest by using
--platform=linux.

Also move the sparse image to image converting process to the ubu
chroot.

Use --no-auto-version switch for building the packages in sfossdk.
Requires updated dhd with this commit:
https://github.com/mer-hybris/droid-hal-device/pull/282/commits/427b5116c557bb3e2676a1ba561a949b4c9cf143